Serverapi/requirements.txt
selenium
beautifulsoup4
mysql-connector-python
webdriver-manager
flask
mysql-connector-python
requests
bcrypt
python-dotenv
pandas